+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # v4.7.0
 ## 🚀 Feature
 - Demoapp deeplink documentation and typography name refactor | Author: [@eportasML](https://github.com/eportasML)
+- `AndesSlider` component | Author: [lenibaldassarre](https://github.com/lenibaldassarre)
 ## 🛠 Fixes
 - AndesFeedbackScreen can be implemented with Flox. | Author: [@orellanaML] (https://github.com/orellanaML)